moved things, created the background
