+++ pages/index.tsx
@@ -1,27 +1,23 @@
 import {
   Box,
-  Button,
   Center,
-  Container,
   Grid,
   GridItem,
   Heading,
-  Input,
   Link,
   Spinner,
   Stack,
   Text,
-  Textarea,
 } from "@chakra-ui/react"
 import isEqual from "lodash/fp/isEqual"
 import type { NextPage } from "next"
 import NextLink from "next/link"
-import React, { useEffect, useState } from "react"
+import { useEffect, useState } from "react"
 import { useDBServiceList } from "../adapters/firebase-database"
 import { NeedDetail, NeedList } from "../components/themed"
-import { Background } from "../components/themed/atoms/Background"
+import { AddNeedModal } from "../components/themed/organisms/AddNeedModal"
 import { useAuth } from "../services/auth"
-import { Need, NeedData, NEED_MODAL_DISPLAY_STATES, UserID } from "../types"
+import { Need, NEED_MODAL_DISPLAY_STATES } from "../types"
 
 const CHURCH_NAME = process.env.NEXT_PUBLIC_CHURCH_NAME
 const Welcome = () => (
@@ -101,146 +97,69 @@ const Home: NextPage = () => {
 
   // RENDER
   return (
-    <Background w="100vw" h="100vh">
-      <Container centerContent maxW="container.xl">
-        {isLoading && <Spinner />}
-        {!currentUser && (
-          <>
-            <Welcome />
-            <Text>Let&apos;s get you logged in.</Text>
-            <NextLink href="/auth" passHref>
-              <Link>Log in</Link>
-            </NextLink>
-          </>
-        )}
-        {loading && <Spinner />}
-        {error && (
-          <Text>Encountered an error loading the Needs data. Sorry!</Text>
-        )}
-        <Stack direction={"column"} spacing={4}>
-          <Welcome />
-          <Grid
-            templateColumns={["1fr", "5fr 1fr"]}
-            gap={6}
-            justifyContent="center"
-            alignItems="center"
-          >
-            <GridItem colSpan={1}>
-              <Box>
-                <NeedList
-                  listSnapshots={[snapshots, loading, error]}
-                  updatedNeedErrorMessage={updatedNeedErrorMessage}
-                  updatedNeedConfirmMessage={updatedNeedConfirmMessage}
-                  setActiveNeed={setActiveNeed}
-                  setShowModal={setShowModal}
-                />
-                {currentUser && showModal === "add" && (
-                  <AddNeedModal
-                    userId={currentUser.id}
-                    handleCloseModal={() => setShowModal("none")}
-                  />
-                )}
-                {currentUser && activeNeed && showModal === "detail" && (
-                  <NeedDetail
-                    userId={currentUser.id}
-                    activeNeed={activeNeed}
-                    userName={currentUser.name}
-                    saveActiveNeed={handleSaveActiveNeed(activeNeed)}
-                    handleCloseDetail={handleCloseDetail}
-                  />
-                )}
-              </Box>
-            </GridItem>
-            <GridItem colStart={2}>
-              <Center w="full">
-                <Box
-                  w={["200px, 100px"]}
-                  h={["200px, 100px"]}
-                  border="1px"
-                  borderColor="teal.200"
-                >
-                  <Text>Enter your QR code here!</Text>
-                </Box>
-              </Center>
-            </GridItem>
-          </Grid>
-        </Stack>
-      </Container>
-    </Background>
-  )
-}
-
-interface AddModalProps {
-  userId: UserID
-  handleCloseModal: () => void
-}
-
-const AddNeedModal: React.FC<AddModalProps> = ({
-  userId,
-  handleCloseModal,
-}) => {
-  const [name, setName] = useState("")
-  const [description, setDescription] = useState("")
-  const [createdNeed, setCreatedNeed] = useState<NeedData | null>(null)
-  const [submitting, setSubmitting] = useState(false)
-
-  useEffect(() => {
-    if (!name || !description || !submitting) {
-      return
-    }
-
-    const post = async () => {
-      const response = await fetch("api/needs", {
-        method: "POST",
-        headers: {
-          Accept: "application/json",
-        },
-        body: JSON.stringify({
-          name,
-          description,
-          fulfilledState: "Unfulfilled",
-          ownerId: userId,
-        }),
-      })
-      const data = await response.json()
-      setCreatedNeed(data.need)
-    }
-
-    post()
-  }, [name, description, userId, submitting])
-
-  useEffect(() => {
-    if (createdNeed) {
-      setSubmitting(false)
-    }
-  }, [createdNeed])
-
-  if (createdNeed && createdNeed.name) {
-    return (
-      <Box>
-        <Text>
-          Created new need: <strong>{createdNeed.name}</strong>
-        </Text>
-        <Button onClick={handleCloseModal}>Close</Button>
-      </Box>
-    )
-  }
-  if (submitting) {
-    return <Spinner />
-  }
-  return (
     <Box>
-      <Input
-        placeholder="Name of your need"
-        value={name}
-        onChange={(e) => setName(e.target.value)}
-      />
-      <Textarea
-        placeholder="A fitting description of your need with appropriate details"
-        value={description}
-        onChange={(e) => setDescription(e.target.value)}
-      />
-      <Button onClick={() => setSubmitting(true)}>Submit</Button>
+      {isLoading && <Spinner />}
+      {!currentUser && (
+        <>
+          <Welcome />
+          <Text>Let&apos;s get you logged in.</Text>
+          <NextLink href="/auth" passHref>
+            <Link>Log in</Link>
+          </NextLink>
+        </>
+      )}
+      {loading && <Spinner />}
+      {error && (
+        <Text>Encountered an error loading the Needs data. Sorry!</Text>
+      )}
+      <Stack direction={"column"} spacing={4}>
+        <Welcome />
+        <Grid
+          templateColumns={["1fr", "5fr 1fr"]}
+          gap={6}
+          justifyContent="center"
+          alignItems="center"
+        >
+          <GridItem colSpan={1}>
+            <Box>
+              <NeedList
+                listSnapshots={[snapshots, loading, error]}
+                updatedNeedErrorMessage={updatedNeedErrorMessage}
+                updatedNeedConfirmMessage={updatedNeedConfirmMessage}
+                setActiveNeed={setActiveNeed}
+                setShowModal={setShowModal}
+              />
+              {currentUser && showModal === "add" && (
+                <AddNeedModal
+                  userId={currentUser.id}
+                  handleCloseModal={() => setShowModal("none")}
+                />
+              )}
+              {currentUser && activeNeed && showModal === "detail" && (
+                <NeedDetail
+                  userId={currentUser.id}
+                  activeNeed={activeNeed}
+                  userName={currentUser.name}
+                  saveActiveNeed={handleSaveActiveNeed(activeNeed)}
+                  handleCloseDetail={handleCloseDetail}
+                />
+              )}
+            </Box>
+          </GridItem>
+          <GridItem colStart={2}>
+            <Center w="full">
+              <Box
+                w={["200px, 100px"]}
+                h={["200px, 100px"]}
+                border="1px"
+                borderColor="teal.200"
+              >
+                <Text>Enter your QR code here!</Text>
+              </Box>
+            </Center>
+          </GridItem>
+        </Grid>
+      </Stack>
     </Box>
   )
 }
